Add files via upload
All main files. Deaths.txt will regenerate itself if removed, however deaths will be 0. Yes.wav is a sample audio file that is meant to be funny, it can be disabled by changing one variable in the python code. Feel free to replace it with an alternative sound file with the same name. SekiroDeath.png is necessary and will not regenerate itself -- if you delete it, you will have to redownload the file.